Created basic voting functionality and restructures itinerary into multiple files
