+++ frontend/src/components/Icons.jsx
@@ -81,4 +81,10 @@ const OtherIcon = () => (
   </svg>
 );
 
-export { CalendarIcon, VoteIcon, MoneyIcon, BasecampIcon, LightIcon, DarkIcon, ProfileIcon, PlusIcon, TrashIcon, FlightIcon, HotelIcon, FoodIcon, ActivityIcon, OtherIcon };
+const CheckIcon = () => (
+  <svg xmlns="http://www.w3.org/2000/svg" fill="black" viewBox="0 0 24 24" strokeWidth={0.2} stroke="black" className="w-5 h-5 mr-2">
+    <path d="M18.7,7.2c-0.4-0.4-1-0.4-1.4,0l-7.5,7.5l-3.1-3.1c0,0,0,0,0,0c-0.4-0.4-1-0.4-1.4,0c-0.4,0.4-0.4,1,0,1.4l3.8,3.8c0.2,0.2,0.4,0.3,0.7,0.3c0.3,0,0.5-0.1,0.7-0.3l8.2-8.2C19.1,8.2,19.1,7.6,18.7,7.2z"/>
+  </svg>
+  );
+
+export { CalendarIcon, VoteIcon, MoneyIcon, BasecampIcon, LightIcon, DarkIcon, ProfileIcon, PlusIcon, TrashIcon, FlightIcon, HotelIcon, FoodIcon, ActivityIcon, OtherIcon, CheckIcon };
